cloud-scanner/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
boto3>=1.28.0
websockets>=11.0.3
python-dotenv>=1.0.0
pytest==7.4.3
pydantic==2.4.2
bcrypt==4.0.1
mangum==0.17.0
email-validator==2.1.0
